requirements.txt
alembic==1.18.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
babel==2.18.0
cuid2==2.0.1
cryptography==47.0.0
Flask==3.1.3
Flask-Alembic==3.2.0
flask-babel==4.0.0
flask-caching==2.4.0
flask-login==0.6.3
flask-sqlalchemy==3.1.1
flask_weasyprint==1.2.0
flask-wtf==1.3.0
loguru==0.7.3
openpyxl==3.1.3
pg8000==1.31.5
pyjwt==2.12.1
PyMySQL==1.1.3
python-ulid==3.1.0
redis==7.4.0
sqlalchemy==2.0.49
teritorio==2025.11.11
waitress==3.0.2
weasyprint==68.1
wtforms==3.2.1
